+++ hnister-bbs-service/src/main/java/cn/zynworld/hnister/hnisterbbsservice/service/BBSThemeService.java
@@ -13,7 +13,7 @@ import java.util.List;
  * @create 2018/4/4
  */
 @Service
-public class BBSThemeService extends BaseAbstractService<BBSTheme,Long,BBSThemeMapper,BBSThemeExample> {
+public class BBSThemeService extends BaseAbstractService<BBSTheme,Integer,BBSThemeMapper,BBSThemeExample> {
 
 
 	/**
